Removes ability to add movies with -ve questions
Adding movies with -ve questions (those for which the answer
is NO) works fine in single player, but can cause problems
in multiplayer. If two players reach the same leaf node
and each adds their movie with a NO question, there's no
sensible way to merge them. So I've removed support for it just now.